Create Panda Knomi Shell for Desk
Whilst the original Design of the shell is suppost to be attached to the Printer, I find it more suitable to place the Knomi on my Desk. As it is connected by WLAN and uses USB for power supply, it can be placed anywhere in the house. I have modified the original P1S shell to accomodate for that. There is a lid with ventilation holes to ensure, that Knomi is not overheating in the housing.